Sommelier: Yet more tidying
