+++ src/main/scala/sommelier/handling/package.scala
@@ -1,0 +1,8 @@
+package sommelier
+
+import sommelier.routing.RequestMatcher
+
+package object handling
+{
+  case class Context(request: Request, matcher: RequestMatcher)
+}
