+++ src/main/scala/sommelier/routing/Matchers.scala
@@ -2,8 +2,6 @@ package sommelier.routing
 
 import java.util.Base64
 
-import sommelier.Rejection
-import sommelier.handling.PathParams
 import sommelier.messaging.{Method, Request, Response}
 
 import scala.util.matching.Regex
