+++ src/main/scala/sommelier/routing/Rejection.scala
@@ -1,0 +1,14 @@
+package sommelier.routing
+
+import sommelier.Response
+
+trait Rejection
+{
+  def response: Response
+}
+object Rejection
+{
+  def apply(r: Response): Rejection = new Rejection {
+    override def response: Response = r
+  }
+}
