+++ src/main/scala/sommelier/routing/Routing.scala
@@ -1,8 +1,8 @@
 package sommelier.routing
 
-import SyncResult.{Accepted, Rejected}
+import sommelier.Context
 import sommelier.messaging.{Method, Request, Response}
-import sommelier.{Context, Rejection, Route}
+import sommelier.routing.SyncResult.{Accepted, Rejected}
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.language.implicitConversions
@@ -24,30 +24,6 @@ object Routing
 
   implicit def statusToResponse(status: Int): Response = Response(status = status)
 
-  implicit class OptionRejection[T](opt: Option[T])
-  {
-    def rejectWith(response: => Response): SyncResult[T] =
-      opt.fold[SyncResult[T]](Rejected(Rejection(response)))(Accepted(_))
-  }
-
-  implicit class AsyncOptionRejection[T](fOpt: Future[Option[T]])
-  {
-    def rejectWith(response: => Response)(implicit executionContext: ExecutionContext): AsyncResult[T] =
-      AsyncResult(fOpt.map(_ rejectWith response))
-  }
-
-  implicit class EitherRejection[L,R](either: Either[L,R])
-  {
-    def rejectWith(ifLeft: L => Response): SyncResult[R] =
-      either.fold[SyncResult[R]](left => Rejected(Rejection(ifLeft(left))), Accepted(_))
-  }
-
-  implicit class AsyncEitherRejection[L,R](fEither: Future[Either[L,R]])
-  {
-    def rejectWith(ifLeft: L => Response)(implicit executionContext: ExecutionContext): AsyncResult[R] =
-      AsyncResult(fEither.map(_ rejectWith ifLeft))
-  }
-
   implicit def responseToResult(response: Response): Result[Response] = Accepted(response)
 
   implicit def responseToRejection(response: Response): Rejection = Rejection(response)
@@ -55,6 +31,8 @@ object Routing
   implicit def requestToResult(request: Request): Result[Request] = Accepted(request)
 
   implicit def statusToResult(status: Int): Result[Response] = Accepted(Response(status))
+
+  implicit def statusToRejection(status: Int): Rejection = Rejection(status)
 
   implicit def futureResponseToResult(fResponse: Future[Response])(implicit executionContext: ExecutionContext): Result[Response] = AsyncResult(fResponse.map(Accepted(_)))
