+++ src/main/scala/sommelier/serving/Router.scala
@@ -4,8 +4,7 @@ import sommelier.events.{Completed, EventBus, ExceptionWhileRouting}
 import sommelier.messaging.{Request, Response}
 import sommelier.routing.ApplyMiddleware.{applyIncoming, applyOutgoing}
 import sommelier.routing.ApplyRoutes.applyRoutes
-import sommelier.routing.{Result, SyncResult}
-import sommelier.{AsyncResult, Middleware, Route, SyncResult}
+import sommelier.routing.{AsyncResult, Middleware, Result, Route, SyncResult}
 
 import scala.concurrent.duration.Duration
 import scala.concurrent.{ExecutionContext, Future}
